created generic abstractions for datatype and converter
