+++ modules/global/src/de/diedavids/cuba/ceuesr/entity/DocumentableConverter.java
@@ -1,70 +1,8 @@
 package de.diedavids.cuba.ceuesr.entity;
 
-import com.google.common.base.Strings;
-import com.haulmont.chile.core.model.MetaClass;
-import com.haulmont.cuba.core.entity.Entity;
-import com.haulmont.cuba.core.global.*;
-
-import javax.persistence.AttributeConverter;
 import javax.persistence.Converter;
 
 @Converter(autoApply = true)
-public class DocumentableConverter implements AttributeConverter<Documentable, String> {
-
-    @Override
-    public String convertToDatabaseColumn(Documentable softReference) {
-
-        if (softReference == null)
-            return "";
-
-        EntityLoadInfoBuilder builder = getEntityLoadInfoBuilder();
-
-        EntityLoadInfo entityLoadInfo = builder.create(softReference);
-
-
-        return entityLoadInfo.toString();
-    }
-
-    @Override
-    public Documentable convertToEntityAttribute(String value) {
-
-
-        if (Strings.isNullOrEmpty(value))
-            return null;
-
-        EntityLoadInfoBuilder builder = getEntityLoadInfoBuilder();
-        EntityLoadInfo entityLoadInfo = builder.parse(value);
-
-        Documentable entity = null;
-
-        if (entityLoadInfo != null) {
-            entity = loadEntity(entityLoadInfo);
-        }
-
-        return entity;
-    }
-
-    private Documentable loadEntity(EntityLoadInfo entityLoadInfo) {
-        DataManager dataManager = getDataManager();
-        return (Documentable) dataManager.load(getLoadContextForForEntityLoadInfo(entityLoadInfo.getMetaClass(), entityLoadInfo.getId()));
-    }
-
-    private DataManager getDataManager() {
-        return AppBeans.get(DataManager.NAME);
-    }
-
-
-    protected LoadContext getLoadContextForForEntityLoadInfo(MetaClass metaClass, Object entityId) {
-        LoadContext loadContext = LoadContext.create(metaClass.getJavaClass());
-        loadContext
-                .setId(entityId);
-        return loadContext;
-    }
-
-
-
-    private EntityLoadInfoBuilder getEntityLoadInfoBuilder() {
-        return AppBeans.get(EntityLoadInfoBuilder.NAME);
-    }
+public class DocumentableConverter extends EntitySoftReferenceInterfaceConverter<Documentable> {
 
 }
